+++ ProjectMER/Features/Serializable/Schematics/SchematicObjectDataList.cs
@@ -4,7 +4,7 @@ public class SchematicObjectDataList
 {
 	public string Path;
 
-	public int RootObjectId { get; set; }
+	public long RootObjectId { get; set; }
 
 	public List<SchematicBlockData> Blocks { get; set; } = new();
 }
